Fix up copy_to_numpy and copy_from_numpy to reflect elwise refactor
Instead of having the dual-mode instantiate as it was, the arrfunc
at the core of these copy operations is now just for one numpy element
to/from one dynd element. This arrfunc is lifted via elwise, and
the setup code for the dynd-like array like "Fixed * Fixed * void"
to represent a 2D NumPy array, for example, is now moved out into
the helper function which calls the arrfunc.