Change the calls to the python bindings arrfuncs
This change removes direct calls to an arrfunc instantiate,
instead going through the standard call mechanism. It is based
on a bit of a hack which wraps the data in a temporary nd::array.

It would be nice to create a skeleton array object for this
purpose, which acts like nd::array but owns no references, relying
on ownership borrowed from others.